Refactor project structure and update memory utilities
Reorganized source files into 'lib' and 'drivers' directories for better modularity and clarity. Replaced custom `memmove` with `memcpy` and introduced `kmemory.{h,c}` for memory utilities. Updated include paths and CMake configuration accordingly to reflect the new structure.
