+++ include/lib/kmemory.h
@@ -1,0 +1,8 @@
+#ifndef KMEMORY_H
+#define KMEMORY_H
+
+#include <stdint.h>
+
+void *memcpy(void *destination, const void *source, uint32_t size);
+
+#endif //KMEMORY_H
