+++ include/lib/kpanic.h
